logi_web/requirements.txt
psycopg2
aio-pika==4.9.3
aiobotocore==0.9.4
aiofiles==0.4.0
aiohttp==3.6.2
aiohttp-apispec==2.2.1
aiohttp-cors==0.7.0
aiohttp-jinja2==1.2.0
aiohttp-rest-framework==0.0.1
aiohttp-security==0.4.0
aiohttp-swagger==1.0.15
aiopg==1.0.0
aioredis==1.1.0
aiounittest==1.4.0
amqplib==1.0.2
apispec==3.3.2
asgiref==3.2.10
async-timeout==3.0.1
asyncpg==0.17.0
attrs==19.3.0
botocore==1.10.58
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cyrusbus==0.1.0
djangorestframework==3.11.1
docutils==0.16
elasticsearch==7.8.1
elasticsearch-async==6.2.0
graypy==1.1.2
hiredis==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jwt==1.0.0
MarkupSafe==1.1.1
marshmallow==3.7.1
multidict==4.7.6
pika==1.1.0
pip-autoremove==0.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-logging-rabbitmq==1.1.1
pytz==2020.1
PyYAML==5.3.1
redis==3.2.1
requests==2.11.1
shortuuid==1.0.1
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.10
webargs==5.5.3
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
gunicorn
weasyprint
numpy
matplotlib